Refactor to current style
